Create 808 Cowbell Bootleg
Might edit the assignment components later, I know there's a better way to do them but this works for now. Also hope to include this as part of a larger functional 808 bootleg (which would have a functional sequencer component and not rely on a while(true) loop).
The paper that I modelled this patch on is here: https://www.researchgate.net/publication/267629988_More_Cowbell_a_Physically-Informed_Circuit-Bendable_Digital_Model_of_the_TR-808_Cowbell
This is just a proof of concept; the ADSR and frequency specifications are copied from the paper but I followed none of the schematics, you can tell it doesn't sound like an actual 808 but that'll come with more processing down the line. For now I like how it sounds.