Create Improved Sine Kick
This one uses trigger logic (just ChucKing a bool of 1 to the envelope trigger for now) instead of a while(true) loop, ADSR means no clicking sound. This is closer to the 808 sound but with no compression (gain is also turned down so there won't be any clipping issues if this patch is used as part of a larger sequencer that triggers multiple instances concurrently).